requirements.txt
aiogram
redis
ujson
uvloop
aiohttp
aioredis
asyncio
cmake
distributed
dlib
ffmpeg-python
gdown
ipython
kornia
lpips
matplotlib
moviepy
ninja
numpy
onnxruntime-gpu
opencv-python
pillow
scipy
tqdm
youtube_dl
aubio
